requirements.txt
flask
flask-bcrypt
flask-pymongo
flask-session
PyJWT
python-dotenv
